Set story mode and invert mode to game mode only vertical/horizontal platforms. Cause no enemies to spawn in start blocks. no enemies in multiplayer.

+++ src/main/java/objects/blocks/platform/Platform.java
@@ -22,7 +22,7 @@ public class Platform extends AGameObject implements IPlatform {
     /**
      * The offSet of the vertical moving platform.
      */
-    private int offSet = 0;
+    private double offSet = 0;
     /**
      * The properties for this platform.
      */
@@ -159,7 +159,7 @@ public class Platform extends AGameObject implements IPlatform {
      * {@inheritDoc}
      */
     @Override
-    public final void setOffset(final int value) {
+    public final void setOffset(final double value) {
         this.offSet = value;
     }
 
@@ -167,7 +167,7 @@ public class Platform extends AGameObject implements IPlatform {
      * {@inheritDoc}
      */
     @Override
-    public final int getOffset() {
+    public final double getOffset() {
         return offSet;
     }
